+++ src/components/SummaryMetric/compositionGroup.js
